pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-agent"
version = "0.1.0"
description = "Forge — Terminal Coding Agent with Harness Engineering"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "anthropic>=0.42.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
    "rich>=13.0",
]

[project.scripts]
forge = "forge.cli:entry"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
